added JWT authentication

+++ app/app.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from fastapi import HTTPException, File, UploadFile, Form, Depends
-from app.schemas import PostCreate, PostResponse
-from app.db import Post, create_db_and_tables, get_async_session
+from app.schemas import PostCreate, PostResponse, UserCreate, UserRead, UserUpdate
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -10,6 +10,7 @@ import shutil
 import os
 import uuid
 import tempfile
+from app.users import auth_backend, current_active_user, fastapi_users
 
 # context manager makes the function have 3 phases: setup, pause, and shutdown
 # this whole function make sure that when server start we have our db and tables ready before some try access them and are safely close once execution finshes
@@ -21,11 +22,18 @@ async def lifespan(app: FastAPI):
 #initializing the fastApi app
 app = FastAPI(lifespan=lifespan)
 
+app.include_router(fastapi_users.get_auth_router(auth_backend), prefix="/auth/jwt", tags=["auth"])
+app.include_router(fastapi_users.get_register_router(UserRead, UserCreate), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_verify_router(UserRead), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix="/users", tags=["Users"])
+
 # UPLOAD endpoint: handles all the file uploads into the databse
 @app.post("/upload")
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(""),
+    user: User = Depends(current_active_user),
     session: AsyncSession = Depends(get_async_session)
 ):
     temp_file_path = None
@@ -44,6 +52,7 @@ async def upload_file(
             )
 
             post = Post(
+                user_id = user.id,
                 caption = caption,
                 url = upload_result.url,
                 file_type = "video" if file.content_type.startswith("video/") else "image",
@@ -66,11 +75,16 @@ async def upload_file(
 #GET Endpoint : returns the uploaded post to the user 
 @app.get("/feed")
 async def get_feed(
-    session: AsyncSession = Depends(get_async_session)
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(current_active_user)
 ):
     # resukt is a cursor object return by the sqlAlchemy which points to the rows in the db and is a set of tuples
     result = await session.execute(select(Post).order_by(Post.created_at.desc())) 
     posts = [row[0] for row in result.all()]  # our main object is at 0th index
+
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
 
     posts_data = []
     
@@ -78,10 +92,13 @@ async def get_feed(
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
+                "email": user_dict.get(post.user_id, "Unknown"),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
+                "is_owner": post.user_id == user.id,
                 "created_at": post.created_at.isoformat()
             }
         )
@@ -90,7 +107,7 @@ async def get_feed(
 
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session : AsyncSession = Depends(get_async_session)):
+async def delete_post(post_id: str, session : AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user)):
     try:
         post_uuid = uuid.UUID(post_id)
 
@@ -100,6 +117,9 @@ async def delete_post(post_id: str, session : AsyncSession = Depends(get_async_s
         if not post:
             raise HTTPException(status_code = 404, detail="Post not found")
         
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You dont have permission to delete this post")
+
         await session.delete(post)
         await session.commit()
